[branch] CacheTagArray constructor's interface changed accordign to new specification: type of argument way has changed to unsigned, and some minor changes
git-svn-id: https://uarch-sim.googlecode.com/svn/branches/2012/borody_task_3@238 8191d7f6-0415-6e15-5cfd-e6cccd3c4e6a

+++ perf_sim/mem/cache_tag_array.cpp
@@ -10,7 +10,7 @@
 #include <cache_tag_array.h>
 
 CacheTagArray::CacheTagArray( unsigned int size_in_bytes,
-                              unsigned short ways,
+                              unsigned int ways,
                               unsigned short block_size_in_bytes, 
                               unsigned short addr_size_in_bits)
 {
